requirements.txt
Flask==3.0.0
gunicorn==21.2.0
flask-cors==4.0.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
numpy==1.24.3
Werkzeug==3.0.1
python-dotenv==1.0.0

